+++ src/util/http/json.go
@@ -5,6 +5,7 @@ package httphelper
 import (
 	"encoding/json"
 	"net/http"
+	"time"
 
 	"github.com/skycoin/skycoin/src/util/logging"
 )
@@ -31,4 +32,40 @@ func SendJSONOr500(log *logging.Logger, w http.ResponseWriter, m interface{}) {
 		log.Errorf("%v", err)
 		Error500(w)
 	}
+}
+
+// Duration JSON type copied from https://github.com/vrischmann/jsonutil, MIT License
+
+// Duration is a wrapper around time.Duration which implements json.Unmarshaler and json.Marshaler.
+// It marshals and unmarshals the duration as a string in the format accepted by time.ParseDuration and returned by time.Duration.String.
+type Duration struct {
+	time.Duration
+}
+
+// FromDuration is a convenience factory to create a Duration instance from the
+// given time.Duration value.
+func FromDuration(d time.Duration) Duration {
+	return Duration{d}
+}
+
+// MarshalJSON implements the json.Marshaler interface. The duration is a quoted-string in the format accepted by time.ParseDuration and returned by time.Duration.String.
+func (d Duration) MarshalJSON() ([]byte, error) {
+	return []byte(`"` + d.String() + `"`), nil
+}
+
+// UnmarshalJSON implements the json.Unmarshaler interface. The duration is expected to be a quoted-string of a duration in the format accepted by time.ParseDuration.
+func (d *Duration) UnmarshalJSON(data []byte) error {
+	var s string
+	if err := json.Unmarshal(data, &s); err != nil {
+		return err
+	}
+
+	tmp, err := time.ParseDuration(s)
+	if err != nil {
+		return err
+	}
+
+	d.Duration = tmp
+
+	return nil
 }